Refactor to remove redux and merge SSR into up-to-date FE work
