--- Comments/ClientApp/src/components/App/App.scss
@@ -1,1 +1,0 @@
-﻿
