--- Comments/ClientApp/src/components/BasicForm/formActions.js
@@ -1,6 +1,0 @@
-export function createComment(comment) {
-	return {
-		type: "CREATE_COMMENT",
-		comment
-	};
-}
