--- Comments/ClientApp/src/components/BasicForm/formReducer.js
@@ -1,8 +1,0 @@
-export default function commentReducer(state = [], action) {
-	switch (action.type) {
-		case "CREATE_COMMENT":
-			return [...state, Object.assign({}, action.comment)];
-		default:
-			return state;
-	}
-}
