--- Comments/ClientApp/src/components/Counter/Counter.js
@@ -1,34 +1,0 @@
-import React, { Component } from "react";
-import { Helmet } from "react-helmet";
-
-export default class Counter extends Component {
-	constructor() {
-		super();
-		this.state = { currentCount: 0 };
-	}
-
-	incrementCounter() {
-		this.setState({
-			currentCount: this.state.currentCount + 1
-		});
-	}
-
-	render() {
-		return (
-			<div>
-				<Helmet>
-					<title>Counter</title>
-				</Helmet>
-				<h1>Counter</h1>
-
-				<p>This is a simple example of a React component.</p>
-
-				<p>
-					Current count: <strong>{this.state.currentCount}</strong>
-				</p>
-
-				<button onClick={() => this.incrementCounter()}>Increment</button>
-			</div>
-		);
-	}
-}
