--- Comments/ClientApp/src/components/FetchData/get_forecasts.js
@@ -1,12 +1,0 @@
-import axios from "axios";
-
-const fetchData = async () => {
-	try {
-		const response = await axios("/api/SampleData/WeatherForecasts/");
-		return response.data;
-	} catch (err) {
-		throw new Error(err);
-	}
-};
-
-export default fetchData;
