+++ Comments/ClientApp/src/main.scss
@@ -2,6 +2,5 @@
 
 @import "nice-design-system";
 
-// Component folder structure imports
-@import "./components/App/App";
-@import "./components/NavMenu/NavMenu";
+// Component folder structure imports here
+
